requirements.txt
streamlit==1.40.0
pillow==11.0.0
